Update access tokens context to call changeset/3 in implementing model
This commit updates the access tokens context such that when creating
a new access token, it first looks for a changeset/3 function in the
implementing model.  If such a function exists, it's called, otherwise,
the AccessGrant module's chagneset/3 function is invoked.

This enables implementations to extend the access grant model to add
support for additional fields in changesets.