Update access token model to accept :refresh_token and :token params
The changeset/3 function in the AccessToken model now accepts the
:refresh_token and :token values as parameters, since that enables
implementation to easily customize what those values look like.

Given the previous change that gives priority to the changeset/3
in the implementation module, rather than the on defined in
the AccessToken module, the addition of these parameters gives
users full control over how tokens are generated.  In fact, it
supersedes the access_token_generator config functionality.